+++ src/tests/integration/api/subraces.itest.js
@@ -16,37 +16,37 @@ afterAll(() => {
   redisClient.quit();
 });
 
-describe('/api/races', () => {
-  it('should list races', async () => {
-    const res = await request(app).get('/api/races');
+describe('/api/subraces', () => {
+  it('should list subraces', async () => {
+    const res = await request(app).get('/api/subraces');
     expect(res.statusCode).toEqual(200);
     expect(res.body.results.length).not.toEqual(0);
   });
 
   describe('with name query', () => {
     it('returns the named object', async () => {
-      const indexRes = await request(app).get('/api/races');
+      const indexRes = await request(app).get('/api/subraces');
       const name = indexRes.body.results[1].name;
-      const res = await request(app).get(`/api/races?name=${name}`);
+      const res = await request(app).get(`/api/subraces?name=${name}`);
       expect(res.statusCode).toEqual(200);
       expect(res.body.results[0].name).toEqual(name);
     });
 
     it('is case insensitive', async () => {
-      const indexRes = await request(app).get('/api/races');
+      const indexRes = await request(app).get('/api/subraces');
       const name = indexRes.body.results[1].name;
       const queryName = name.toLowerCase();
-      const res = await request(app).get(`/api/races?name=${queryName}`);
+      const res = await request(app).get(`/api/subraces?name=${queryName}`);
       expect(res.statusCode).toEqual(200);
       expect(res.body.results[0].name).toEqual(name);
     });
   });
 
-  describe('/api/races/:index', () => {
+  describe('/api/subraces/:index', () => {
     it('should return one object', async () => {
-      const indexRes = await request(app).get('/api/races');
+      const indexRes = await request(app).get('/api/subraces');
       const index = indexRes.body.results[0].index;
-      const showRes = await request(app).get(`/api/races/${index}`);
+      const showRes = await request(app).get(`/api/subraces/${index}`);
       expect(showRes.statusCode).toEqual(200);
       expect(showRes.body.index).toEqual(index);
     });
@@ -54,36 +54,24 @@ describe('/api/races', () => {
     describe('with an invalid index', () => {
       it('should return one object', async () => {
         const invalidIndex = 'invalid-index';
-        const showRes = await request(app).get(`/api/races/${invalidIndex}`);
+        const showRes = await request(app).get(`/api/subraces/${invalidIndex}`);
         expect(showRes.statusCode).toEqual(404);
       });
     });
 
-    describe('/api/races/:index/subraces', () => {
+    describe('/api/subraces/:index/traits', () => {
       it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
+        const indexRes = await request(app).get('/api/subraces');
         const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/subraces`);
+        const res = await request(app).get(`/api/subraces/${index}/traits`);
         expect(res.statusCode).toEqual(200);
         expect(res.body.results.length).not.toEqual(0);
       });
     });
 
-    describe('/api/races/:index/proficiencies', () => {
+    describe('/api/subraces/:index/proficiencies', () => {
       it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
-        const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/proficiencies`);
-        expect(res.statusCode).toEqual(200);
-        expect(res.body.results.length).not.toEqual(0);
-      });
-    });
-
-    describe('/api/races/:index/traits', () => {
-      it('returns objects', async () => {
-        const indexRes = await request(app).get('/api/races');
-        const index = indexRes.body.results[1].index;
-        const res = await request(app).get(`/api/races/${index}/traits`);
+        const res = await request(app).get(`/api/subraces/high-elf/proficiencies`);
         expect(res.statusCode).toEqual(200);
         expect(res.body.results.length).not.toEqual(0);
       });
